refactor: move XPU SLA adapter to platform layer
