+++ lightx2v_platform/ops/attn/intel_xpu/xpu_sla_attn.py
@@ -33,3 +33,44 @@ def sparse_block_attention(q, k, v, lut, block_q=128, block_k=128, scale=None):
 def sla_sparse_attention(q, k, v, keep_ratio=0.2, block_q=128, block_k=128, scale=None):
     """Route and apply sparse attention through ``sycl_kernels.sla_sparse_attention``."""
     return _get_sycl_api("sla_sparse_attention")(q, k, v, keep_ratio, block_q, block_k, scale)
+
+
+def apply_intel_xpu_cute_attn(
+    q,
+    k,
+    v,
+    cu_seqlens_q=None,
+    cu_seqlens_kv=None,
+    max_seqlen_q=None,
+    max_seqlen_kv=None,
+    *,
+    keep_ratio=0.2,
+    block_q=128,
+    block_k=128,
+    **kwargs,
+):
+    """Run the XPU SLA router and fused sparse attention on one sequence."""
+    if q.ndim != 3 or k.ndim != 3 or v.ndim != 3:
+        raise ValueError("Intel XPU SLA expects q, k and v in [L, H, D] layout")
+    if q.shape != k.shape or k.shape != v.shape:
+        raise ValueError("Intel XPU SLA currently requires self-attention with matching q/k/v shapes")
+    if max_seqlen_q is None:
+        max_seqlen_q = q.shape[0]
+    if max_seqlen_kv is None:
+        max_seqlen_kv = k.shape[0]
+    if max_seqlen_q != q.shape[0] or max_seqlen_kv != k.shape[0]:
+        raise ValueError("Intel XPU SLA currently supports one unpadded sequence per call")
+    if cu_seqlens_q is not None and cu_seqlens_q.numel() != 2:
+        raise ValueError("Intel XPU SLA currently supports one sequence per call")
+    if cu_seqlens_kv is not None and cu_seqlens_kv.numel() != 2:
+        raise ValueError("Intel XPU SLA currently supports one sequence per call")
+
+    out = sla_sparse_attention(
+        q.unsqueeze(0),
+        k.unsqueeze(0),
+        v.unsqueeze(0),
+        keep_ratio=keep_ratio,
+        block_q=block_q,
+        block_k=block_k,
+    )
+    return out.reshape(max_seqlen_q, -1)
